Improve fragment environment for validity, add stop_mask & original fragment stems (#52)
Changes the fragment environment slightly so that the agent is forced to produce valid molecules. This is done by further masking invalid actions, in particular, the agent is no longer allowed to stop unless inter-fragment bonds are labeled. 

Also further constrains the agent's choice of bonding site by using the (more restrictive) original stem attachment points of the fragments of [Bengio et al. (2021)](https://arxiv.org/abs/2106.04399). 

This means the fragment environment will be slightly less expressive, but closer to the original paper and slightly easier to train.

* improve fragment environment for validity + include stop_mask + include original fragment stems

* tox

* remove temp code

* Remove json prettifier pre-commit hook

* fix mask to device in GraphActionCategorical

* more comments and slightly better architecture adaptation

* replace JSON file of fragments with constant in bengio2021flow